chore: add modularized scaffold (feature-based clean architecture)
