+++ .github/workflows/ci.yaml
@@ -20,6 +20,12 @@ jobs:
             - name: Install dependencies
               run: go mod download
 
+            - name: Run golangci-lint
+              uses: golangci/golangci-lint-action@v6
+              with:
+                  version: latest
+                  args: --timeout=5m
+
             - name: Run build
               run: go build ./...
